fix(ios-tests): install console proxy at app bootstrap so it survives reloads
Previous run showed Tap hidden uncle as 'proxy installed: false' — the page
reloads mid-test on that one, wiping the WDIO-installed window proxy along
with the new window object. Move the proxy into the app itself so every page
load (including hard reloads) reinstalls it via initialize.ts.

- New src/util/iOSConsoleProxy.ts: gated by ?__ios_console_proxy URL flag, no-op otherwise. Module-scope logs buffer; drainiOSConsoleLogs() is the canonical reader (atomic splice). install hangs references on window.__iOSConsoleProxy__ (peek) and window.__drainiOSConsoleLogs__ (cross-realm drain).
- initialize.ts calls installiOSConsoleProxy() at module load, before initialize() runs.
- wdio.base.conf.ts removes the old browser.execute install block; before hook navigates to '?__ios_console_proxy' so the URL flag persists across browser.refresh() in beforeTest. afterTest drains via window.__drainiOSConsoleLogs__.
